backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Data validation
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0

# Database
asyncpg==0.30.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.4.0
httpx==0.28.1
itsdangerous==2.2.0

# AI
google-generativeai==0.8.3

# Utilities
python-dotenv==1.0.1
starlette==0.41.3

# Dapr (optional — only used in Kubernetes)
dapr==1.14.0
